пока просто перенес тест (изначальный) из кода в библиотеку с тестами
